Allow decimal targetHours and targetHoursByDay in 15-minute increments
Settings targets are now decimal, so values like 4.25, 6.5, or 7.75 work
alongside whole hours. Values that aren't a multiple of 0.25 are invalid:
a per-day override falls back to the flat targetHours, and an invalid or
non-positive targetHours falls back to the default 8 (previously 4.25
failed int deserialization and silently reset all settings to defaults).
The status bar formats the target invariantly ("62% of 4.25h").

Closes #35

+++ src/Punch.CLI/PunchSession.cs
@@ -7,7 +7,7 @@ namespace Punch.CLI;
 // The event loop mutates these in place and the renderer reads them.
 internal sealed class PunchSession
 {
-    public PunchSession(DaySchedule schedule, DateOnly workingDate, string filePath, int cursorSlot, int targetHours = 8)
+    public PunchSession(DaySchedule schedule, DateOnly workingDate, string filePath, int cursorSlot, decimal targetHours = 8)
     {
         Schedule = schedule;
         WorkingDate = workingDate;
@@ -20,8 +20,9 @@ internal sealed class PunchSession
     public DateOnly WorkingDate { get; }
     public string FilePath { get; }
 
-    // The daily workday goal in whole hours, used for the status-bar percentage.
-    public int TargetHours { get; }
+    // The daily workday goal in hours (15-minute increments), used for the
+    // status-bar percentage.
+    public decimal TargetHours { get; }
 
     public IReadOnlyList<TimeBlock> Blocks => Schedule.Blocks;
